fix the conflict with payment
